Flight Management Application Deployment for Airline Company Project/GA/backend/requirements.txt
flask
flask-mysql-connector
flask-cors
